linux: kselftests: Install headers into B directory
A recent commit in the Linux kernel makes a great deal out of
"tainted" kernel sources when building out of that directory.
In OE's case, the objects are always built elsewhere, not in
the kernel source tree itself, so this case applies.

While that is fine for most of the kernel tasks, the append
that we have for kselftests means that we run `make
headers_install`, which, as it seems, end up in the kernel
source tree.

Setting KBUILD_OUTPUT before actually installing the headers
helps in maintaining that source tree clean.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>