kselftests: Add Perl as a run-time dependency
A recent commit ("selftests: Move test output to diagnostic
lines") introduced a Perl script that takes care of changing
the output of tests to comments, so that all tests are TAP13-
compliant.

Perl is generally in all images, except in X15 where we reduce
the number of package due to partition size. This adds Perl
explicitly.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>